[feat] 메인페이지 commonlist와 contentdetail 페이지 useparams와 라우터 링크로 연결 구현. / 기존 test 삭제 후 원본에 적용.
